Rewrite the low level networking stuff
The main motivator behind this was that the old API exposed the
IPv4/IPv6 separation to user code, which then proceeded to get it wrong.
The new API eliminates that class of bugs entirely.

The other motivator was that we need to send separate IPv4/IPv6 master
server heartbeats, which didn't really need a complete rewrite but would
have been annoying.