Better fix + explanation for d0ab92f7cd00d "WTF"
`constexpr char asdf[] = "hello";` seems to just ignore the constexpr
requirement and reconstructs the string at runtime. In debug builds it
would then return to-be-corrupted stack data, in release builds it got
nuked for being UB.

`constexpr const char *` embeds the literal as a readonly constant as
you would expect.