Nuke democams
This code has been a consistent drag since we forked. The code is awful,
nobody has tested or used it at any point in the last five years, and
it gets disproportionally caught up in any refactoring we do.

There is no way the code still works, the only reason we didn't kill it
already is because some of the regular spectator/freefly code runs
through it.