Add dieselfont tool
This is a rewrite of dieselfont (github.com/mikejsavage/dieselfont) to live in
the repo and not depend on boost with all the settings living in uncommitted
files and/or shell history on my PC, which had been making it borderline
unusable.

Note that you need to uncomment dieselfont in make.lua because the msdfgen
dependency makes it quite slow to compile. Also note the half complete DIY
implementation in dieselfont.cpp.

This commit also changes the font metadata spec to allow us to do stuff like
baseline alignment (needed for cinematic_mapname ents) properly.