Add MP3 support
MP3 at VBR0 is smaller than the oggs we have now and decode nearly 3x as
fast.

MP3s click when looping, so we keep the old oggs and stb_vorbis around
for looping sounds. There is no dev advantage between MP3/ogg like with
PNG/DDS so we arbitrarily prefer ogg over MP3 when hotloading and warn
on collisions.